SubscribeCoffieBackend/analytics/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1
supabase==2.3.1
